+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,15 +1,18 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 516
+INVENTREE_API_VERSION = 517
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
 
-v516 -> 2026-07-03 : https://github.com/inventree/InvenTree/pull/11971
+v517 -> 2026-07-06 : https://github.com/inventree/InvenTree/pull/11971
     - Removes direct "notes" field from any models which previously supported markdown notes
     - Adds a generic "Note" model which can be attached to any model type via a generic foreign key relationship
     - Allow multiple notes to be attached to a single object, and for notes to be created / edited / deleted via the API
+
+v516 -> 2026-07-03 : https://github.com/inventree/InvenTree/pull/12295
+    - Adds "consumable" field to the Part model and API endpoints
 
 v515 -> 2026-07-03 : https://github.com/inventree/InvenTree/pull/12298
     - Change the file fields definition to binary (from uri) in the upload requests